renaming navPagination to pageRender
